fix(replay): satisfy package boundary gates
